refactor the minimap to not use state topojson (#2349)
* refactor the minimap to not use state topojson
* remove redundant component and memoize some calls
