--- src/components/CountyMap/countyTopoJson/CT.json
@@ -1,1 +1,0 @@
-{"type":"Topology","transform":{"scale":[0.001765973231065926,0.001245282128958586],"translate":[-73.72800337305297,40.9851110032158]},"arcs":[[[714,481],[32,14]],[[746,495],[47,3],[50,87]],[[843,585],[46,-45],[6,8],[48,-12],[61,-14],[55,7],[39,-3]],[[1098,526],[0,-35],[0,0],[-2,-41],[-3,-103],[-24,-4],[2,-47],[-14,-27],[-14,13],[-40,-5],[-37,-11],[-41,-2],[-23,-12],[-38,13],[-19,-12],[-32,-16],[-27,-2]],[[786,235],[-20,79],[-33,38],[72,10],[-9,66],[-61,5],[-21,48]],[[553,477],[21,20],[0,19],[96,12],[22,4],[22,-51]],[[786,235],[-5,0],[-21,-13],[-11,14],[-38,-7],[-35,-13]],[[676,216],[-45,60],[-23,88],[-14,-4],[-38,-8],[1,60],[-3,37],[-1,28]],[[119,547],[7,126],[10,182],[145,-4],[59,-2],[42,-2],[25,-1]],[[407,846],[-11,-58],[80,5],[-11,-41],[-24,-92],[-38,-7],[10,-69],[9,-58]],[[422,526],[-43,-27],[-60,-40],[5,-33],[-27,-6],[-28,3],[-37,-5],[4,-30]],[[236,388],[-74,18],[-37,141],[-6,0]],[[422,526],[25,3],[-4,-70],[57,-10],[53,28]],[[676,216],[-7,-3],[-29,15],[-32,-3],[0,0],[-20,-15],[-40,-4],[-15,19],[-54,-18],[-12,5],[-18,8],[-20,-14],[-8,-6],[-1,-1],[-19,-21],[-34,-10],[-12,-25]],[[355,143],[-12,13],[20,60],[0,18],[11,20],[-27,25],[-18,28],[-21,20],[-13,23],[-51,22],[-8,16]],[[407,846],[5,0],[87,-1],[21,-31],[20,3],[4,0],[18,27],[72,-4],[45,3],[11,-1]],[[690,842],[9,-69],[-11,-2],[-2,-60],[11,-8],[19,-92],[31,-93],[-1,-23]],[[843,585],[47,32],[-35,66],[1,98],[37,1],[29,0],[-1,56]],[[921,838],[65,-1],[58,-2],[47,-1],[1,-13],[4,-161],[2,-66],[0,0],[0,-68]],[[690,842],[63,-1],[46,0],[67,-2],[36,-1],[19,0]],[[104,306],[1,7],[0,1],[0,1],[3,51],[4,69],[6,92],[1,20]],[[355,143],[-4,-8],[-13,-5],[-24,5],[-17,4],[-33,-33],[-5,0],[-15,-3],[-19,-3],[-24,-5],[11,-14],[-19,-22],[-20,-9],[-26,3],[-27,-10],[-26,-18],[-19,0],[-35,-25],[0,0],[-1,15],[0,11],[-39,67],[18,12],[121,78],[-39,66],[4,57]]],"objects":{"cb_2015_connecticut_county_20m":{"type":"GeometryCollection","geometries":[{"arcs":[[0,1,2,3,4]],"type":"Polygon","properties":{"STATEFP":"09","COUNTYFP":"011","COUNTYNS":"00212799","AFFGEOID":"0500000US09011","GEOID":"09011","NAME":"New London","LSAD":"06","ALAND":1722641876,"AWATER":276619712}},{"arcs":[[5,-5,6,7]],"type":"Polygon","properties":{"STATEFP":"09","COUNTYFP":"007","COUNTYNS":"00212797","AFFGEOID":"0500000US09007","GEOID":"09007","NAME":"Middlesex","LSAD":"06","ALAND":956491412,"AWATER":180679144}},{"arcs":[[8,9,10,11]],"type":"Polygon","properties":{"STATEFP":"09","COUNTYFP":"005","COUNTYNS":"00212796","AFFGEOID":"0500000US09005","GEOID":"09005","NAME":"Litchfield","LSAD":"06","ALAND":2384262297,"AWATER":62138320}},{"arcs":[[12,-8,13,14,-11]],"type":"Polygon","properties":{"STATEFP":"09","COUNTYFP":"009","COUNTYNS":"00212798","AFFGEOID":"0500000US09009","GEOID":"09009","NAME":"New Haven","LSAD":"06","ALAND":1565645638,"AWATER":667096946}},{"arcs":[[15,16,-1,-6,-13,-10]],"type":"Polygon","properties":{"STATEFP":"09","COUNTYFP":"003","COUNTYNS":"00212338","AFFGEOID":"0500000US09003","GEOID":"09003","NAME":"Hartford","LSAD":"06","ALAND":1903947639,"AWATER":40133958}},{"arcs":[[17,18,-3]],"type":"Polygon","properties":{"STATEFP":"09","COUNTYFP":"015","COUNTYNS":"00212801","AFFGEOID":"0500000US09015","GEOID":"09015","NAME":"Windham","LSAD":"06","ALAND":1328185576,"AWATER":21767064}},{"arcs":[[19,-18,-2,-17]],"type":"Polygon","properties":{"STATEFP":"09","COUNTYFP":"013","COUNTYNS":"00212668","AFFGEOID":"0500000US09013","GEOID":"09013","NAME":"Tolland","LSAD":"06","ALAND":1062613133,"AWATER":17533745}},{"arcs":[[20,-12,-15,21]],"type":"Polygon","properties":{"STATEFP":"09","COUNTYFP":"001","COUNTYNS":"00212794","AFFGEOID":"0500000US09001","GEOID":"09001","NAME":"Fairfield","LSAD":"06","ALAND":1618664455,"AWATER":548955210}}]}}}
