--- src/components/CountyMap/countyTopoJson/DE.json
@@ -1,1 +1,0 @@
-{"type":"Topology","transform":{"scale":[0.0017659732310659098,0.0012452821289585868],"translate":[-75.78889413370692,38.45096187078507]},"arcs":[[[37,304],[95,5],[41,55],[41,10],[17,36],[42,-11]],[[273,399],[2,-7],[0,-21],[40,-55],[42,-44],[26,-5],[13,11],[10,-81],[10,-128],[3,-69],[-77,0],[-89,1],[-78,1],[-121,5],[-4,67],[-1,14],[-3,60],[0,0],[-9,156]],[[110,1111],[14,4],[50,-8],[38,-22],[-15,-17],[-10,-12],[-11,-41],[-18,-23],[-14,-31],[-14,-14],[9,-27],[17,-15],[-8,-34],[0,-30],[-37,-15],[12,-33],[28,-41],[9,-14]],[[160,738],[-44,-49],[-20,1],[-18,-15],[-36,6],[-26,-2]],[[16,679],[-4,65],[0,0],[-12,277],[9,0],[11,29],[21,27],[30,24],[39,10]],[[160,738],[21,-31],[34,-54],[8,-61],[-7,-44],[6,-61],[32,-30],[19,-58]],[[37,304],[-14,252],[-5,83],[-2,40]]],"objects":{"cb_2015_delaware_county_20m":{"type":"GeometryCollection","geometries":[{"arcs":[[0,1]],"type":"Polygon","properties":{"STATEFP":"10","COUNTYFP":"005","COUNTYNS":"00217269","AFFGEOID":"0500000US10005","GEOID":"10005","NAME":"Sussex","LSAD":"06","ALAND":2424465375,"AWATER":674255133}},{"arcs":[[2,3,4]],"type":"Polygon","properties":{"STATEFP":"10","COUNTYFP":"003","COUNTYNS":"00217270","AFFGEOID":"0500000US10003","GEOID":"10003","NAME":"New Castle","LSAD":"06","ALAND":1104074671,"AWATER":175385786}},{"arcs":[[-4,5,-1,6]],"type":"Polygon","properties":{"STATEFP":"10","COUNTYFP":"001","COUNTYNS":"00217271","AFFGEOID":"0500000US10001","GEOID":"10001","NAME":"Kent","LSAD":"06","ALAND":1518565081,"AWATER":549102079}}]}}}
