--- src/components/CountyMap/countyTopoJson/HI.json
@@ -1,1 +1,0 @@
-{"type":"Topology","transform":{"scale":[0.0017671214581342482,0.0012462184313111915],"translate":[-160.25009585589743,18.91746639594062]},"arcs":[[[1831,1833],[17,9],[12,0],[24,-35],[2,0]],[[1886,1807],[10,-30],[-10,-2],[-14,25],[-18,12],[-23,6],[0,15]],[[1142,2139],[60,3],[26,33],[16,34],[33,30],[14,4],[25,-50],[27,-61],[22,-36],[-4,-51],[18,-21],[27,17],[24,-2],[-1,-44],[8,-36],[33,-36],[-2,-3],[-10,-20],[-15,-17],[-32,12],[-13,-11],[-17,-6],[-24,22],[-22,17],[-34,5],[-17,3],[-25,-5],[-36,-9],[-8,0],[-15,37],[-1,4],[-5,20],[-22,23],[-2,21],[-29,46],[1,29],[-26,44],[26,4]],[[264,2525],[22,27],[8,34],[14,15],[54,34],[17,18],[21,-2],[19,-14],[13,21],[31,-8],[17,10],[23,-15],[28,-25],[11,-48],[-14,-49],[-10,-16],[2,-34],[-3,-39],[-27,-31],[-34,-35],[-46,12],[-27,7],[-17,0],[-26,33],[-33,22],[-27,14],[-18,32],[2,37]],[[0,2325],[12,59],[20,27],[32,21],[8,12],[6,26],[23,6],[7,-6],[5,-12],[-11,-14],[-9,-29],[4,-25],[-26,-16],[-18,-7],[-10,-17],[-9,-19],[-9,-35],[-14,8],[-11,21]],[[2377,692],[24,30],[17,26],[15,6],[19,24],[14,36],[20,29],[14,11],[3,35],[-14,30],[-22,49],[-7,43],[7,62],[21,13],[31,-14],[35,-25],[18,-25],[38,-31],[22,-23],[22,6],[32,-14],[65,-38],[67,-42],[59,-62],[23,-32],[22,-34],[-3,-63],[3,-38],[23,9],[22,0],[15,-39],[3,-46],[16,-23],[53,-45],[22,-15],[-1,-24],[-34,-54],[-38,-41],[-44,-41],[-52,-32],[-26,-18],[-27,-6],[-33,10],[-28,-18],[-26,-34],[-17,-6],[-36,-40],[-29,-11],[-29,-55],[-20,-50],[-13,-29],[-13,-24],[-20,-19],[-30,42],[-46,35],[-42,19],[-19,50],[1,64],[6,64],[6,32],[2,40],[-12,54],[-9,31],[-15,26],[-10,56],[-5,42],[-11,27],[-18,6],[-2,16],[-18,49],[9,39]],[[2009,1608],[11,47],[35,39],[32,-10],[25,-49],[21,-45],[44,14],[40,24],[51,-7],[27,-36],[35,-25],[41,-41],[32,-12],[10,-41],[-9,-37],[-24,-27],[-49,-30],[-44,1],[-42,-26],[-53,-14],[-31,16],[-7,46],[-8,65],[-2,14],[-6,29],[-19,7],[-18,-17],[-9,7],[-44,28],[-32,56],[-7,24]],[[2026,1318],[34,27],[25,9],[13,-20],[2,-42],[-26,-13],[-47,-5],[-19,22],[18,22]],[[1806,1602],[28,13],[41,-4],[36,-24],[21,-25],[16,-35],[-17,-45],[-29,-16],[-11,-4],[-33,-3],[-13,32],[-1,41],[-11,22],[-27,25],[0,23]],[[1886,1807],[43,-1],[56,7],[19,-14],[-17,-38],[-36,-36],[-42,-14],[-43,13],[-41,19],[-31,10],[-52,-9],[-46,-2],[-33,11],[19,46],[16,20],[-6,34],[33,-6],[42,-14],[50,-8],[14,8]]],"objects":{"cb_2015_hawaii_county_20m":{"type":"GeometryCollection","geometries":[{"arcs":[[0,1]],"type":"Polygon","properties":{"STATEFP":"15","COUNTYFP":"005","COUNTYNS":"01702380","AFFGEOID":"0500000US15005","GEOID":"15005","NAME":"Kalawao","LSAD":"06","ALAND":31057602,"AWATER":105769486}},{"arcs":[[2]],"type":"Polygon","properties":{"STATEFP":"15","COUNTYFP":"003","COUNTYNS":"00365281","AFFGEOID":"0500000US15003","GEOID":"15003","NAME":"Honolulu","LSAD":"06","ALAND":1555505156,"AWATER":4054940027}},{"arcs":[[[3]],[[4]]],"type":"MultiPolygon","properties":{"STATEFP":"15","COUNTYFP":"007","COUNTYNS":"00365282","AFFGEOID":"0500000US15007","GEOID":"15007","NAME":"Kauai","LSAD":"06","ALAND":1605560469,"AWATER":1674243566}},{"arcs":[[5]],"type":"Polygon","properties":{"STATEFP":"15","COUNTYFP":"001","COUNTYNS":"00365280","AFFGEOID":"0500000US15001","GEOID":"15001","NAME":"Hawaii","LSAD":"06","ALAND":10433651107,"AWATER":2739477691}},{"arcs":[[[6]],[[7]],[[8]],[[-2,9]]],"type":"MultiPolygon","properties":{"STATEFP":"15","COUNTYFP":"009","COUNTYNS":"00365283","AFFGEOID":"0500000US15009","GEOID":"15009","NAME":"Maui","LSAD":"06","ALAND":3008359138,"AWATER":3203293646}}]}}}
