--- src/components/CountyMap/countyTopoJson/MA.json
@@ -1,1 +1,0 @@
-{"type":"Topology","transform":{"scale":[0.001766844456142932,0.0012462248103582512],"translate":[-73.5090226924008,41.2379032753944]},"arcs":[[[328,1204],[37,-1],[31,-1],[166,-6],[33,-1],[4,0],[22,-1],[73,-3]],[[694,1191],[6,-38],[27,-29],[-29,-50],[3,-24],[15,-27],[-27,-27],[-13,-109]],[[676,887],[-23,-32],[-11,94],[-61,-11],[-4,21],[-101,-19],[-18,35],[-33,-6],[-64,31],[-3,46],[-56,11]],[[302,1057],[14,69],[1,50],[-42,30]],[[275,1206],[3,0],[50,-2]],[[288,862],[27,25],[38,-9],[-15,-75],[67,-2],[61,-43],[41,83],[3,-1],[5,-14],[-14,-27],[9,-14],[10,-4],[81,13],[30,-34],[34,28],[24,9],[40,11]],[[729,808],[-24,-49],[73,-18],[0,-106]],[[778,635],[-36,1],[-67,2],[-46,0],[-63,1],[-11,1],[-45,-3],[-72,4],[-18,-27],[-4,0],[-20,-3],[-21,31],[-87,1],[-5,0],[-25,1]],[[258,644],[-10,86],[36,72],[3,16],[1,44]],[[1374,688],[75,42],[13,8],[6,105],[-19,-5]],[[1449,838],[17,18],[21,-4],[17,-26],[9,-2]],[[1513,824],[-10,-21],[7,-21],[8,-10],[23,20],[2,19]],[[1543,811],[29,-31],[26,-62],[3,-6],[23,-36],[-3,-27],[-19,-33],[9,-36],[31,-16],[14,8],[21,-27],[11,-46],[-8,-35],[2,-3]],[[1682,461],[-16,-20],[-38,-19],[6,-13],[-39,-70],[-14,-37]],[[1581,302],[-26,21],[-2,1],[-22,-32]],[[1531,292],[-20,11],[-15,68],[-11,45],[-13,6],[-8,22],[-59,-10],[7,17],[-13,14],[36,35],[-8,36],[-3,1],[-4,18],[-31,45],[-15,88]],[[676,887],[15,7],[8,-41],[34,-6],[2,-34],[-6,-5]],[[288,862],[-35,13],[-2,48],[32,-1],[17,89],[2,46]],[[1513,824],[26,-9],[4,-4]],[[1149,737],[76,35],[23,38],[-15,52],[34,12],[22,-5],[23,-31]],[[1312,838],[1,-2],[24,-20],[9,-22],[4,6],[-2,3],[10,7],[0,6],[-2,3],[0,1],[6,0],[5,5],[23,9],[6,22],[28,6]],[[1424,862],[14,-36],[11,12]],[[1374,688],[-33,-19],[-84,-45],[0,0],[-37,-20],[-16,-4]],[[1204,600],[0,27],[-44,-1],[-22,-1]],[[1138,625],[0,38],[-1,31],[12,23],[0,20]],[[1312,838],[19,38],[-6,24],[28,3],[23,-4],[7,8],[-9,11],[56,30]],[[1430,948],[4,-51],[-1,-1],[-4,-9],[-8,-18],[3,-7]],[[1514,93],[43,10],[40,60],[47,33],[29,-23],[32,-23],[18,-4],[9,-53],[-73,-4],[-66,-5],[-42,-31],[-30,-42],[-7,52],[0,30]],[[1575,245],[16,-13],[16,-11]],[[1607,221],[-37,-22],[-31,-32],[-39,-16],[-51,-13],[8,36],[72,35],[46,36]],[[89,1021],[25,98],[24,91],[70,-2],[67,-2]],[[258,644],[-42,2],[-59,2],[-145,4],[-1,3],[-11,25],[56,213],[15,60],[18,68]],[[1830,58],[46,3],[65,5],[17,58],[37,-27],[14,-76],[-32,-21],[-46,2],[-65,7],[-36,49]],[[694,1191],[45,-2],[45,-2],[24,-1],[87,-3],[17,-1]],[[912,1182],[23,-62],[46,8],[62,-26],[-8,-65],[25,-5],[55,15],[-3,-61],[-46,-94],[22,-31],[20,12],[36,3],[-11,-52],[-9,1],[-15,0],[-21,-6],[17,-61],[30,7],[14,-28]],[[1138,625],[-35,-1],[-18,-1],[-9,0],[-108,-5],[-1,13],[-47,1],[-58,2],[-65,1],[-19,0]],[[1275,1202],[6,0],[36,1],[-3,43],[21,20],[19,-3],[29,-5],[9,20],[10,23],[37,8],[20,12],[37,-11],[27,1],[7,-72],[19,-57],[24,-33],[27,-6],[21,22],[24,-10],[5,-34],[-35,-42],[-25,-4],[-59,-13],[-25,-9],[7,-48],[-29,-16],[-7,-17],[-4,-8],[-4,-10],[-24,-1],[-16,-2]],[[1429,951],[-40,42],[12,88],[-64,19],[-11,-9],[-51,47],[10,11],[-10,53]],[[912,1182],[53,-2],[33,-1],[61,-2],[162,-5],[32,-1],[22,31],[0,0]],[[1429,951],[1,-3]],[[1531,292],[-10,-15],[-18,4],[-2,-1],[-30,-7],[0,-1],[-18,-29],[-23,-24],[-22,-7],[-8,-2],[-29,8],[-20,-10],[-9,85],[3,46],[-36,12],[-3,5],[-4,7],[-30,50],[-32,20],[-7,4],[-6,40],[1,49],[-24,0],[0,48],[0,26]],[[1682,461],[24,-31],[30,-16],[67,-14],[29,-12],[32,18],[53,12],[55,23],[12,17],[-1,35],[-36,27],[-3,43],[-1,15],[-4,44],[-38,11],[-23,-5],[-31,34],[32,15],[28,9],[51,-23],[31,-52],[8,-42],[3,-17],[22,-93],[5,-94],[0,-1],[-2,-55],[-19,-57],[-17,-4],[-9,-2],[-1,61],[3,41],[-27,-5],[-59,-12],[-63,-26],[-29,10],[-45,-19],[-25,-12],[-18,-27],[-47,-8],[-53,-23],[-9,-5]],[[1575,245],[16,13],[1,35],[-11,9]]],"objects":{"cb_2015_massachusetts_county_20m":{"type":"GeometryCollection","geometries":[{"arcs":[[0,1,2,3,4]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"011","COUNTYNS":"00606932","AFFGEOID":"0500000US25011","GEOID":"25011","NAME":"Franklin","LSAD":"06","ALAND":1811271672,"AWATER":65368074}},{"arcs":[[5,6,7,8]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"013","COUNTYNS":"00606933","AFFGEOID":"0500000US25013","GEOID":"25013","NAME":"Hampden","LSAD":"06","ALAND":1598409204,"AWATER":44293472}},{"arcs":[[9,10,11,12,13,14,15]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"023","COUNTYNS":"00606938","AFFGEOID":"0500000US25023","GEOID":"25023","NAME":"Plymouth","LSAD":"06","ALAND":1707122280,"AWATER":1124387947}},{"arcs":[[-3,16,-6,17]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"015","COUNTYNS":"00606934","AFFGEOID":"0500000US25015","GEOID":"25015","NAME":"Hampshire","LSAD":"06","ALAND":1365598019,"AWATER":46487557}},{"arcs":[[[18,-12]],[[19,20,21,-10,22,23,24]]],"type":"MultiPolygon","properties":{"STATEFP":"25","COUNTYFP":"021","COUNTYNS":"00606937","AFFGEOID":"0500000US25021","GEOID":"25021","NAME":"Norfolk","LSAD":"06","ALAND":1025927532,"AWATER":124925285}},{"arcs":[[25,26,-21]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"025","COUNTYNS":"00606939","AFFGEOID":"0500000US25025","GEOID":"25025","NAME":"Suffolk","LSAD":"06","ALAND":150881016,"AWATER":160487602}},{"arcs":[[[27]],[[28,29]]],"type":"MultiPolygon","properties":{"STATEFP":"25","COUNTYFP":"007","COUNTYNS":"00606930","AFFGEOID":"0500000US25007","GEOID":"25007","NAME":"Dukes","LSAD":"06","ALAND":267455894,"AWATER":1004127633}},{"arcs":[[30,-4,-18,-9,31]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"003","COUNTYNS":"00606928","AFFGEOID":"0500000US25003","GEOID":"25003","NAME":"Berkshire","LSAD":"06","ALAND":2400491048,"AWATER":50746697}},{"arcs":[[32]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"019","COUNTYNS":"00606936","AFFGEOID":"0500000US25019","GEOID":"25019","NAME":"Nantucket","LSAD":"06","ALAND":119431671,"AWATER":667032212}},{"arcs":[[33,34,-25,35,-7,-17,-2]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"027","COUNTYNS":"00606940","AFFGEOID":"0500000US25027","GEOID":"25027","NAME":"Worcester","LSAD":"06","ALAND":3912698515,"AWATER":177395253}},{"arcs":[[36,37]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"009","COUNTYNS":"00606931","AFFGEOID":"0500000US25009","GEOID":"25009","NAME":"Essex","LSAD":"06","ALAND":1275669707,"AWATER":870068134}},{"arcs":[[38,-38,39,-26,-20,-35]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"017","COUNTYNS":"00606935","AFFGEOID":"0500000US25017","GEOID":"25017","NAME":"Middlesex","LSAD":"06","ALAND":2118147279,"AWATER":75727649}},{"arcs":[[-23,-16,40]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"005","COUNTYNS":"00606929","AFFGEOID":"0500000US25005","GEOID":"25005","NAME":"Bristol","LSAD":"06","ALAND":1432602435,"AWATER":357662687}},{"arcs":[[-14,41,-29,42]],"type":"Polygon","properties":{"STATEFP":"25","COUNTYFP":"001","COUNTYNS":"00606927","AFFGEOID":"0500000US25001","GEOID":"25001","NAME":"Barnstable","LSAD":"06","ALAND":1019719638,"AWATER":2361606119}}]}}}
