--- src/components/CountyMap/countyTopoJson/NH.json
@@ -1,1 +1,0 @@
-{"type":"Topology","transform":{"scale":[0.0017659732310659401,0.0012458771036977022],"translate":[-72.5553971476252,42.697373930533864]},"arcs":[[[579,857],[53,-55],[54,-53],[0,0],[0,-1],[0,0],[47,-46],[55,-26]],[[788,676],[2,-147],[-43,-57]],[[747,472],[-131,100],[-82,23],[-65,101]],[[469,696],[43,46],[2,52],[37,5],[34,4],[-6,54]],[[293,346],[1,43],[15,5]],[[309,394],[55,15],[58,-66],[44,12],[37,11],[15,-79],[69,11],[3,-16],[69,-32]],[[659,250],[8,-49],[-42,-14],[6,-33],[8,-52],[37,-20],[0,-13],[66,-38]],[[742,31],[-6,0],[0,0],[-22,-31],[-32,1],[-162,5],[-62,2],[-33,1],[-53,2],[-17,1]],[[355,12],[-26,62],[-15,104],[-35,21],[3,59],[11,73],[0,15]],[[645,1215],[45,-101],[48,127],[119,-5],[18,38]],[[875,1274],[1,-21],[4,-133],[6,-203],[2,-37],[3,-74],[5,-105],[5,-24]],[[901,677],[-34,-46],[-7,37],[-72,8]],[[579,857],[-6,25],[-10,64],[113,32],[-11,146],[-20,91]],[[449,1369],[41,14],[25,39],[37,26],[-4,37],[24,25],[-4,39],[-18,30],[-24,66],[32,47],[27,60],[16,25],[-21,52],[17,30],[2,45],[28,32],[16,49],[8,22],[-19,33],[45,25],[43,26],[22,-28],[35,-21],[25,1],[17,32],[14,18],[15,-245],[12,-212],[8,-189],[2,-48],[3,-80],[1,-32],[1,-13]],[[645,1215],[-15,30],[-75,5],[-7,40],[-62,22],[-37,57]],[[351,667],[25,37],[15,20],[27,-44],[51,16]],[[747,472],[-6,-9]],[[741,463],[-30,-70],[-10,-19],[-22,-49],[-27,-49],[7,-26]],[[309,394],[-45,108],[87,165]],[[69,430],[7,25],[11,45],[-7,34],[10,53],[0,1],[9,51],[1,65],[26,25]],[[126,729],[128,-42],[9,-19],[12,19],[76,-20]],[[293,346],[-80,3],[13,40],[-60,-10],[-96,-27],[-11,21]],[[59,373],[6,43],[4,14]],[[741,463],[67,-59],[58,-50],[-27,-45],[112,-1],[32,34]],[[983,342],[36,-35],[29,-16],[-17,-38],[-17,-30],[-19,-41],[-7,-28],[-4,-14],[-27,-1],[-37,11],[-20,-12],[-37,-8],[-10,-23],[-9,-20],[-29,5],[-19,3],[-21,-20],[3,-43],[-36,-1]],[[901,677],[5,-25],[-3,-28],[-4,-36],[-9,-43],[34,-41],[29,-44],[34,-43],[-7,-30],[0,0],[-2,-40],[5,-5]],[[195,864],[16,28],[7,52],[37,61],[-7,35],[21,29],[0,0],[3,4],[22,55],[-10,45],[-4,21],[4,49],[4,37],[25,26],[33,10],[39,0],[21,8],[11,4],[23,38],[9,3]],[[126,729],[2,3],[25,89],[36,32],[6,11]],[[355,12],[-87,3],[-24,1],[-45,2],[-45,2],[-73,3],[-22,1],[-4,0],[-11,27],[-35,35],[-8,43],[-1,7],[14,25],[-1,46],[23,10],[27,30],[-4,12],[-6,20],[15,34],[-9,53],[0,7]]],"objects":{"cb_2015_new_hampshire_county_20m":{"type":"GeometryCollection","geometries":[{"arcs":[[0,1,2,3]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"001","COUNTYNS":"00873174","AFFGEOID":"0500000US33001","GEOID":"33001","NAME":"Belknap","LSAD":"06","ALAND":1036582285,"AWATER":177039289}},{"arcs":[[4,5,6,7,8]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"011","COUNTYNS":"00873179","AFFGEOID":"0500000US33011","GEOID":"33011","NAME":"Hillsborough","LSAD":"06","ALAND":2269346872,"AWATER":41478107}},{"arcs":[[9,10,11,-1,12]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"003","COUNTYNS":"00873175","AFFGEOID":"0500000US33003","GEOID":"33003","NAME":"Carroll","LSAD":"06","ALAND":2411467696,"AWATER":158924712}},{"arcs":[[13,-10,14]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"007","COUNTYNS":"00873177","AFFGEOID":"0500000US33007","GEOID":"33007","NAME":"Coos","LSAD":"06","ALAND":4648216850,"AWATER":90773914}},{"arcs":[[15,-3,16,17,-6,18]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"013","COUNTYNS":"00873180","AFFGEOID":"0500000US33013","GEOID":"33013","NAME":"Merrimack","LSAD":"06","ALAND":2419397118,"AWATER":57788836}},{"arcs":[[19,20,-19,-5,21,22]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"019","COUNTYNS":"00873183","AFFGEOID":"0500000US33019","GEOID":"33019","NAME":"Sullivan","LSAD":"06","ALAND":1391587914,"AWATER":38070571}},{"arcs":[[-18,23,24,-7]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"015","COUNTYNS":"00873181","AFFGEOID":"0500000US33015","GEOID":"33015","NAME":"Rockingham","LSAD":"06","ALAND":1799732526,"AWATER":259591042}},{"arcs":[[-2,-12,25,-24,-17]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"017","COUNTYNS":"00873182","AFFGEOID":"0500000US33017","GEOID":"33017","NAME":"Strafford","LSAD":"06","ALAND":955401999,"AWATER":39157533}},{"arcs":[[26,-15,-13,-4,-16,-21,27]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"009","COUNTYNS":"00873178","AFFGEOID":"0500000US33009","GEOID":"33009","NAME":"Grafton","LSAD":"06","ALAND":4425926973,"AWATER":105375540}},{"arcs":[[-22,-9,28]],"type":"Polygon","properties":{"STATEFP":"33","COUNTYFP":"005","COUNTYNS":"00873176","AFFGEOID":"0500000US33005","GEOID":"33005","NAME":"Cheshire","LSAD":"06","ALAND":1830365875,"AWATER":57990894}}]}}}
