--- src/components/CountyMap/countyTopoJson/NV.json
@@ -1,1 +1,0 @@
-{"type":"Topology","transform":{"scale":[0.0017670194237383696,0.001246169082041894],"translate":[-120.0053318931358,35.00153037356978]},"arcs":[[[1935,4814],[243,0],[0,-268],[89,-433],[94,1]],[[2361,4114],[-1,-89],[23,-132],[-21,-99],[11,-109],[-36,-103],[-17,-243]],[[2320,3339],[-393,0]],[[1927,3339],[4,171],[-4,559],[4,199],[-2,503],[6,15],[0,28]],[[2,3299],[136,1],[119,-22]],[[257,3278],[136,-2],[1,-52],[-52,-28],[-22,-82],[51,-123],[12,-156]],[[383,2835],[-145,144],[-1,1],[-180,175],[-55,54],[0,54],[0,36]],[[394,4782],[296,1],[0,-85],[196,0],[450,2],[0,-140],[194,-1],[1,-126]],[[1531,4433],[-137,-421]],[[1394,4012],[-543,-2],[-409,0]],[[442,4010],[-65,0],[15,421],[2,351]],[[409,3707],[-111,-227],[-95,-57],[-37,-13]],[[166,3410],[30,65],[-21,150],[42,-6],[88,58],[104,30]],[[2325,1477],[89,9],[553,-8],[403,0]],[[3370,1478],[1,-192],[1,-224],[-1,-45],[1,-60],[-30,-59],[-22,-55],[-7,-22],[-35,-8],[-33,16],[-25,22],[-12,37],[-20,28],[-26,2],[-17,-16],[-22,3],[-14,17],[-34,1],[-32,-9],[-22,-19],[-39,-10],[-4,-28],[0,-3],[0,-45],[7,-53],[17,-34],[17,-29],[-19,-40],[4,-65],[4,-66],[21,-33],[-6,-69],[-1,-60],[21,-32],[23,-119],[8,-99],[-15,-14],[-11,0],[-3,-42],[-5,-54],[-97,111],[-135,154],[-66,74],[-82,92],[-57,64],[-136,154],[-1,0],[-112,123]],[[2354,772],[0,127],[-28,40],[-1,538]],[[2832,2950],[539,0]],[[3371,2950],[-1,-84],[0,0],[0,-135],[0,-124],[0,-82],[0,0],[0,-214],[-1,-109],[0,-113],[0,0],[0,-90],[0,-167],[0,-157],[1,-71],[0,-126]],[[2325,1477],[2,969],[505,1],[0,503]],[[0,3463],[1,102],[1,76],[1,147],[1,188],[1,136],[0,157],[-1,321],[-1,116],[0,255],[1,349],[-1,206],[0,96],[155,1],[207,-2],[21,0]],[[386,5611],[0,-465],[10,-145],[-13,-208],[11,-11]],[[442,4010],[9,-152],[22,-17],[0,-87],[-11,-23]],[[462,3731],[-39,-2],[-14,-22]],[[166,3410],[-33,-28],[-42,-12],[-20,-28],[-70,0]],[[1,3342],[-1,101],[0,20]],[[166,3410],[36,-44],[49,-1],[6,-87]],[[2,3299],[-1,43]],[[1310,2786],[297,-379],[0,-826]],[[1607,1581],[-45,47],[-247,260],[-86,89],[-107,111],[-230,235]],[[892,2323],[44,-2],[374,465]],[[1394,4012],[45,-92],[-19,-111],[19,-101],[-13,-62],[-59,-23],[-38,-39],[6,-34],[-64,-45],[22,-125],[-49,-54],[18,-42]],[[1262,3284],[-51,-16]],[[1211,3268],[-502,1]],[[709,3269],[0,30],[-207,221],[16,104],[17,178],[-73,-71]],[[386,5611],[65,0],[118,0],[127,-1],[155,2],[172,2],[183,1],[142,0],[124,0],[117,1],[97,0],[4,0]],[[1690,5616],[0,-802]],[[1690,4814],[0,-287],[-129,0],[-30,-94]],[[480,2739],[142,0],[3,351],[-64,-1],[0,74],[48,104],[15,1],[85,1]],[[1211,3268],[-188,-59],[0,-65],[287,-358]],[[892,2323],[-41,42],[-253,257],[-118,117]],[[2361,4114],[322,-5],[573,-4],[116,0]],[[3372,4105],[0,-15],[0,-154],[0,-90],[0,-202],[0,-35],[-1,-396],[0,-102],[0,-161]],[[2832,2950],[-512,389]],[[480,2739],[-69,68],[-28,28]],[[1262,3284],[251,56],[414,-1]],[[2354,772],[-26,29],[-114,125],[-160,174],[-63,69],[-291,313],[-93,99]],[[1690,4814],[245,0]],[[1690,5616],[223,-2],[166,0],[262,-1],[138,1],[176,-1],[156,2],[78,1],[171,-4],[179,-1],[136,0],[1,-193],[0,-210],[-1,-228],[0,-166],[0,0],[-1,-220],[-1,-185],[-1,-78],[0,-226]]],"objects":{"cb_2015_nevada_county_20m":{"type":"GeometryCollection","geometries":[{"arcs":[[0,1,2,3]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"011","COUNTYNS":"00858640","AFFGEOID":"0500000US32011","GEOID":"32011","NAME":"Eureka","LSAD":"06","ALAND":10814963933,"AWATER":11082875}},{"arcs":[[4,5,6]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"005","COUNTYNS":"00858643","AFFGEOID":"0500000US32005","GEOID":"32005","NAME":"Douglas","LSAD":"06","ALAND":1838162296,"AWATER":72317570}},{"arcs":[[7,8,9,10]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"027","COUNTYNS":"00858646","AFFGEOID":"0500000US32027","GEOID":"32027","NAME":"Pershing","LSAD":"06","ALAND":15634723159,"AWATER":79934451}},{"arcs":[[11,12]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"029","COUNTYNS":"00858642","AFFGEOID":"0500000US32029","GEOID":"32029","NAME":"Storey","LSAD":"06","ALAND":680962190,"AWATER":1871089}},{"arcs":[[13,14,15]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"003","COUNTYNS":"00863600","AFFGEOID":"0500000US32003","GEOID":"32003","NAME":"Clark","LSAD":"06","ALAND":20438834319,"AWATER":438841388}},{"arcs":[[16,17,-14,18]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"017","COUNTYNS":"00863992","AFFGEOID":"0500000US32017","GEOID":"32017","NAME":"Lincoln","LSAD":"06","ALAND":27540364862,"AWATER":9802713}},{"arcs":[[19,20,-11,21,22,-13,23,24]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"031","COUNTYNS":"00858608","AFFGEOID":"0500000US32031","GEOID":"32031","NAME":"Washoe","LSAD":"06","ALAND":16323089034,"AWATER":621162127}},{"arcs":[[-24,25,-5,26]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"510","COUNTYNS":"00863219","AFFGEOID":"0500000US32510","GEOID":"32510","NAME":"Carson City","LSAD":"00","ALAND":374669606,"AWATER":32587484}},{"arcs":[[27,28,29]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"009","COUNTYNS":"00858641","AFFGEOID":"0500000US32009","GEOID":"32009","NAME":"Esmeralda","LSAD":"06","ALAND":9277052939,"AWATER":18552011}},{"arcs":[[-10,30,31,32,33,-22]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"001","COUNTYNS":"00858638","AFFGEOID":"0500000US32001","GEOID":"32001","NAME":"Churchill","LSAD":"06","ALAND":12770148576,"AWATER":242405502}},{"arcs":[[34,35,36,-8,-21]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"013","COUNTYNS":"00863310","AFFGEOID":"0500000US32013","GEOID":"32013","NAME":"Humboldt","LSAD":"06","ALAND":24969524844,"AWATER":44196979}},{"arcs":[[37,-33,38,-30,39]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"021","COUNTYNS":"00858644","AFFGEOID":"0500000US32021","GEOID":"32021","NAME":"Mineral","LSAD":"06","ALAND":9719802099,"AWATER":155746685}},{"arcs":[[40,41,-17,42,-2]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"033","COUNTYNS":"00858615","AFFGEOID":"0500000US32033","GEOID":"32033","NAME":"White Pine","LSAD":"06","ALAND":22985054689,"AWATER":54678877}},{"arcs":[[-12,-23,-34,-38,43,-6,-26]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"019","COUNTYNS":"00858645","AFFGEOID":"0500000US32019","GEOID":"32019","NAME":"Lyon","LSAD":"06","ALAND":5183310353,"AWATER":59602530}},{"arcs":[[-32,44,-3,-43,-19,-16,45,-28,-39]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"023","COUNTYNS":"00863599","AFFGEOID":"0500000US32023","GEOID":"32023","NAME":"Nye","LSAD":"06","ALAND":47091002405,"AWATER":43207981}},{"arcs":[[-9,-37,46,-4,-45,-31]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"015","COUNTYNS":"00863368","AFFGEOID":"0500000US32015","GEOID":"32015","NAME":"Lander","LSAD":"06","ALAND":14219255502,"AWATER":76103226}},{"arcs":[[47,-41,-1,-47,-36]],"type":"Polygon","properties":{"STATEFP":"32","COUNTYFP":"007","COUNTYNS":"00857664","AFFGEOID":"0500000US32007","GEOID":"32007","NAME":"Elko","LSAD":"06","ALAND":44471193293,"AWATER":85747701}}]}}}
