--- src/components/CountyMap/countyTopoJson/RI.json
@@ -1,1 +1,0 @@
-{"type":"Topology","transform":{"scale":[0.0017659732310659274,0.0012452821289585824],"translate":[-71.86313564104735,41.14575239785147]},"arcs":[[[42,465],[189,6],[32,29],[12,-11],[7,-15]],[[282,474],[19,-39]],[[301,435],[-28,-30]],[[273,405],[-59,-38],[-172,-5]],[[42,362],[0,0],[0,35],[0,68]],[[301,435],[42,-40],[28,42]],[[371,437],[4,-7],[3,-5],[36,-12],[-3,-46],[9,-85],[-11,-9],[-30,-22],[-11,7],[-15,9],[-4,2],[-22,6],[-15,-30],[-22,0],[-22,8],[-15,-4]],[[253,249],[5,18],[15,138]],[[131,17],[22,57],[27,4],[6,-49],[9,-26],[-42,-3],[-22,17]],[[1,140],[14,27],[-2,47],[24,4],[3,103],[2,41]],[[253,249],[-7,-1],[-15,-37],[-16,-30],[-41,2],[-39,-10],[-44,-19],[-15,-3],[-32,-6],[-44,-13],[1,8]],[[36,692],[109,5],[9,0],[18,1],[35,1],[22,1],[44,1],[0,-27],[0,-26],[0,-48],[24,0],[-1,-49],[6,-40],[7,-4]],[[309,507],[-27,-33]],[[42,465],[0,0],[-2,66],[-4,161]],[[309,507],[32,-20],[30,-50]]],"objects":{"cb_2015_rhode_island_county_20m":{"type":"GeometryCollection","geometries":[{"arcs":[[0,1,2,3,4]],"type":"Polygon","properties":{"STATEFP":"44","COUNTYFP":"003","COUNTYNS":"01219778","AFFGEOID":"0500000US44003","GEOID":"44003","NAME":"Kent","LSAD":"06","ALAND":436515567,"AWATER":50720026}},{"arcs":[[-3,5,6,7]],"type":"Polygon","properties":{"STATEFP":"44","COUNTYFP":"005","COUNTYNS":"01219779","AFFGEOID":"0500000US44005","GEOID":"44005","NAME":"Newport","LSAD":"06","ALAND":265220482,"AWATER":546983257}},{"arcs":[[[8]],[[9,-4,-8,10]]],"type":"MultiPolygon","properties":{"STATEFP":"44","COUNTYFP":"009","COUNTYNS":"01219782","AFFGEOID":"0500000US44009","GEOID":"44009","NAME":"Washington","LSAD":"06","ALAND":852834829,"AWATER":604715929}},{"arcs":[[11,12,-1,13]],"type":"Polygon","properties":{"STATEFP":"44","COUNTYFP":"007","COUNTYNS":"01219781","AFFGEOID":"0500000US44007","GEOID":"44007","NAME":"Providence","LSAD":"06","ALAND":1060637931,"AWATER":67704263}},{"arcs":[[-13,14,-6,-2]],"type":"Polygon","properties":{"STATEFP":"44","COUNTYFP":"001","COUNTYNS":"01219777","AFFGEOID":"0500000US44001","GEOID":"44001","NAME":"Bristol","LSAD":"06","ALAND":62550804,"AWATER":53350670}}]}}}
