--- src/components/CountyMap/countyTopoJson/VT.json
@@ -1,1 +1,0 @@
-{"type":"Topology","transform":{"scale":[0.0017659732310659395,0.0012459603871987643],"translate":[-73.4366177899271,42.72726070162887]},"arcs":[[[53,1833],[85,1]],[[138,1834],[-27,-49],[6,-19],[21,-15],[-36,-120],[16,-29]],[[118,1602],[-21,-105],[-55,-24]],[[42,1473],[-3,4],[-13,42],[14,64],[5,41],[13,41],[-4,5],[-14,15],[-8,24],[23,49],[-3,42],[1,33]],[[350,424],[75,-26],[143,8]],[[568,406],[-4,-14],[-6,-43],[0,-7],[9,-53],[-15,-34],[6,-20],[4,-12],[-27,-30],[-23,-10],[1,-46],[-14,-25],[1,-7],[8,-43],[35,-35],[11,-27],[-33,1],[-166,6],[-31,1],[-37,1]],[[287,9],[5,151],[-49,3],[3,74],[3,78],[73,-5],[11,120],[17,-6]],[[92,174],[2,70],[0,4],[9,224]],[[103,472],[158,-19],[61,8]],[[322,461],[27,-2],[1,-35]],[[287,9],[-50,2],[-3,0],[-67,2],[-70,2],[-15,45],[7,25],[3,89]],[[67,1233],[195,24],[5,-51],[7,-55]],[[274,1151],[11,-6],[14,-66],[17,-3],[77,-31]],[[393,1045],[-28,-55]],[[365,990],[-41,-79],[-7,32],[-59,-16],[13,-44]],[[271,883],[-33,-18],[-8,34],[-107,-8],[7,-56],[-87,-11]],[[43,824],[6,14],[-18,30],[-5,7],[9,46],[-19,44],[-2,36],[3,29],[-17,25],[11,46],[10,42],[2,12],[26,51],[19,22],[-1,5]],[[871,1831],[117,2],[46,1],[62,1],[-17,-30],[21,-52],[-16,-25],[-27,-60],[-32,-47],[24,-66],[18,-30],[4,-39],[-24,-25],[4,-37],[-37,-26],[-25,-39],[-41,-14],[-9,-3],[-23,-38],[-11,-4]],[[905,1300],[-54,75],[44,44],[-30,41],[38,52],[-39,29],[42,56],[-58,42]],[[848,1639],[-20,14],[57,72],[-31,25],[17,81]],[[601,1490],[64,-33],[84,166],[57,-41],[42,57]],[[905,1300],[-21,-8],[-39,0],[-33,-10],[-25,-26],[-4,-37],[-4,-49],[4,-21]],[[783,1149],[-66,26],[-76,-7],[-36,17]],[[605,1185],[29,76],[27,33],[30,57],[-2,8],[-2,3],[-55,31],[-64,34]],[[568,1427],[33,63]],[[118,1602],[108,-58],[64,-15]],[[290,1529],[39,-44],[26,-24],[-16,-71],[19,-6]],[[358,1384],[-20,-39],[8,-36],[-17,6],[-32,-68],[9,-4],[-15,-43],[-17,-49]],[[67,1233],[-3,37],[-6,38],[7,20],[16,47],[-11,53],[-20,32],[-8,13]],[[500,1832],[12,-2],[104,-1],[184,1],[62,0],[9,1]],[[601,1490],[-61,40],[37,72],[-91,47]],[[486,1649],[26,41],[-12,142]],[[365,990],[5,-26],[118,-45],[206,-79]],[[694,840],[-6,-11],[-36,-32],[-25,-89],[-2,-3],[-26,-25],[-1,-65],[-9,-51],[0,-1],[-10,-53],[7,-34],[-11,-45],[-7,-25]],[[322,461],[8,56],[53,-14],[21,87],[-27,13],[-11,14],[10,77],[8,-3],[33,76],[-68,28],[1,23],[20,47],[-8,20],[-43,8],[-5,13],[-43,-23]],[[393,1045],[33,-14],[31,61],[16,37],[7,21],[83,-26],[42,61]],[[783,1149],[10,-45],[-22,-55],[-3,-4],[0,0],[-21,-30],[7,-35],[-37,-61],[-7,-51],[-16,-28]],[[103,472],[1,39],[6,138],[-28,39],[-20,34],[-39,-47],[-16,25],[5,48],[12,32],[19,44]],[[290,1529],[42,63],[31,-15],[5,28],[22,47],[96,-3]],[[568,1427],[-34,-66],[-58,45],[-33,-65],[-63,35],[-22,8]],[[138,1834],[13,1],[69,1],[63,0],[201,-3],[16,-1]]],"objects":{"cb_2015_vermont_county_20m":{"type":"GeometryCollection","geometries":[{"arcs":[[0,1,2,3]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"013","COUNTYNS":"01461763","AFFGEOID":"0500000US50013","GEOID":"50013","NAME":"Grand Isle","LSAD":"06","ALAND":211894597,"AWATER":292303559}},{"arcs":[[4,5,6]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"025","COUNTYNS":"01461769","AFFGEOID":"0500000US50025","GEOID":"50025","NAME":"Windham","LSAD":"06","ALAND":2033922469,"AWATER":33064491}},{"arcs":[[7,8,9,-7,10]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"003","COUNTYNS":"01461758","AFFGEOID":"0500000US50003","GEOID":"50003","NAME":"Bennington","LSAD":"06","ALAND":1748198169,"AWATER":6926826}},{"arcs":[[11,12,13,14,15,16]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"001","COUNTYNS":"01461757","AFFGEOID":"0500000US50001","GEOID":"50001","NAME":"Addison","LSAD":"06","ALAND":1984779894,"AWATER":107303299}},{"arcs":[[17,18,19]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"009","COUNTYNS":"01461761","AFFGEOID":"0500000US50009","GEOID":"50009","NAME":"Essex","LSAD":"06","ALAND":1718707340,"AWATER":30285614}},{"arcs":[[20,-19,21,22,23,24]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"005","COUNTYNS":"01461759","AFFGEOID":"0500000US50005","GEOID":"50005","NAME":"Caledonia","LSAD":"06","ALAND":1680879242,"AWATER":22456093}},{"arcs":[[-3,25,26,27,-12,28]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"007","COUNTYNS":"01461760","AFFGEOID":"0500000US50007","GEOID":"50007","NAME":"Chittenden","LSAD":"06","ALAND":1389740227,"AWATER":214237435}},{"arcs":[[29,-20,-21,30,31]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"019","COUNTYNS":"01461766","AFFGEOID":"0500000US50019","GEOID":"50019","NAME":"Orleans","LSAD":"06","ALAND":1796376593,"AWATER":71442197}},{"arcs":[[32,33,-5,-10,34,-15]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"027","COUNTYNS":"01461770","AFFGEOID":"0500000US50027","GEOID":"50027","NAME":"Windsor","LSAD":"06","ALAND":2510577932,"AWATER":19279313}},{"arcs":[[-14,35,-23,36,-33]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"017","COUNTYNS":"01461765","AFFGEOID":"0500000US50017","GEOID":"50017","NAME":"Orange","LSAD":"06","ALAND":1779409624,"AWATER":13549371}},{"arcs":[[-16,-35,-9,37]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"021","COUNTYNS":"01461767","AFFGEOID":"0500000US50021","GEOID":"50021","NAME":"Rutland","LSAD":"06","ALAND":2408211831,"AWATER":38758591}},{"arcs":[[38,-31,-25,39,-27]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"015","COUNTYNS":"01461764","AFFGEOID":"0500000US50015","GEOID":"50015","NAME":"Lamoille","LSAD":"06","ALAND":1188473311,"AWATER":12658077}},{"arcs":[[-28,-40,-24,-36,-13]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"023","COUNTYNS":"01461768","AFFGEOID":"0500000US50023","GEOID":"50023","NAME":"Washington","LSAD":"06","ALAND":1779419614,"AWATER":21187618}},{"arcs":[[40,-32,-39,-26,-2]],"type":"Polygon","properties":{"STATEFP":"50","COUNTYFP":"011","COUNTYNS":"01461762","AFFGEOID":"0500000US50011","GEOID":"50011","NAME":"Franklin","LSAD":"06","ALAND":1641633750,"AWATER":150930318}}]}}}
